float32->float64

+++ ease/easing_functions.go
@@ -5,28 +5,28 @@ import (
 	"math"
 )
 
-const backS float32 = 1.70158
+const backS float64 = 1.70158
 
-var pi = float32(math.Pi)
+var pi = float64(math.Pi)
 
 // TweenFunc provides an interface used for the easing equation. You can use
 // one of the provided easing functions or provide your own.
-type TweenFunc func(t, b, c, d float32) float32
+type TweenFunc func(t, b, c, d float64) float64
 
-func Linear(t, b, c, d float32) float32 {
+func Linear(t, b, c, d float64) float64 {
 	return c*t/d + b
 }
 
-func InQuad(t, b, c, d float32) float32 {
+func InQuad(t, b, c, d float64) float64 {
 	return c*pow(t/d, 2) + b
 }
 
-func OutQuad(t, b, c, d float32) float32 {
+func OutQuad(t, b, c, d float64) float64 {
 	t = t / d
 	return -c*t*(t-2) + b
 }
 
-func InOutQuad(t, b, c, d float32) float32 {
+func InOutQuad(t, b, c, d float64) float64 {
 	t = t / d * 2
 	if t < 1 {
 		return c/2*pow(t, 2) + b
@@ -34,22 +34,22 @@ func InOutQuad(t, b, c, d float32) float32 {
 	return -c/2*((t-1)*(t-3)-1) + b
 }
 
-func OutInQuad(t, b, c, d float32) float32 {
+func OutInQuad(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutQuad(t*2, b, c/2, d)
 	}
 	return InQuad((t*2)-d, b+c/2, c/2, d)
 }
 
-func InCubic(t, b, c, d float32) float32 {
+func InCubic(t, b, c, d float64) float64 {
 	return c*pow(t/d, 3) + b
 }
 
-func OutCubic(t, b, c, d float32) float32 {
+func OutCubic(t, b, c, d float64) float64 {
 	return c*(pow(t/d-1, 3)+1) + b
 }
 
-func InOutCubic(t, b, c, d float32) float32 {
+func InOutCubic(t, b, c, d float64) float64 {
 	t = t / d * 2
 	if t < 1 {
 		return c/2*t*t*t + b
@@ -58,22 +58,22 @@ func InOutCubic(t, b, c, d float32) float32 {
 	return c/2*(t*t*t+2) + b
 }
 
-func OutInCubic(t, b, c, d float32) float32 {
+func OutInCubic(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutCubic(t*2, b, c/2, d)
 	}
 	return InCubic((t*2)-d, b+c/2, c/2, d)
 }
 
-func InQuart(t, b, c, d float32) float32 {
+func InQuart(t, b, c, d float64) float64 {
 	return c*pow(t/d, 4) + b
 }
 
-func OutQuart(t, b, c, d float32) float32 {
+func OutQuart(t, b, c, d float64) float64 {
 	return -c*(pow(t/d-1, 4)-1) + b
 }
 
-func InOutQuart(t, b, c, d float32) float32 {
+func InOutQuart(t, b, c, d float64) float64 {
 	t = t / d * 2
 	if t < 1 {
 		return c/2*pow(t, 4) + b
@@ -81,22 +81,22 @@ func InOutQuart(t, b, c, d float32) float32 {
 	return -c/2*(pow(t-2, 4)-2) + b
 }
 
-func OutInQuart(t, b, c, d float32) float32 {
+func OutInQuart(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutQuart(t*2, b, c/2, d)
 	}
 	return InQuart((t*2)-d, b+c/2, c/2, d)
 }
 
-func InQuint(t, b, c, d float32) float32 {
+func InQuint(t, b, c, d float64) float64 {
 	return c*pow(t/d, 5) + b
 }
 
-func OutQuint(t, b, c, d float32) float32 {
+func OutQuint(t, b, c, d float64) float64 {
 	return c*(pow(t/d-1, 5)+1) + b
 }
 
-func InOutQuint(t, b, c, d float32) float32 {
+func InOutQuint(t, b, c, d float64) float64 {
 	t = t / d * 2
 	if t < 1 {
 		return c/2*pow(t, 5) + b
@@ -104,47 +104,47 @@ func InOutQuint(t, b, c, d float32) float32 {
 	return c/2*(pow(t-2, 5)+2) + b
 }
 
-func OutInQuint(t, b, c, d float32) float32 {
+func OutInQuint(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutQuint(t*2, b, c/2, d)
 	}
 	return InQuint((t*2)-d, b+c/2, c/2, d)
 }
 
-func InSine(t, b, c, d float32) float32 {
+func InSine(t, b, c, d float64) float64 {
 	return -c*cos(t/d*(pi/2)) + c + b
 }
 
-func OutSine(t, b, c, d float32) float32 {
+func OutSine(t, b, c, d float64) float64 {
 	return c*sin(t/d*(pi/2)) + b
 }
 
-func InOutSine(t, b, c, d float32) float32 {
+func InOutSine(t, b, c, d float64) float64 {
 	return -c/2*(cos(pi*t/d)-1) + b
 }
 
-func OutInSine(t, b, c, d float32) float32 {
+func OutInSine(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutSine(t*2, b, c/2, d)
 	}
 	return InSine((t*2)-d, b+c/2, c/2, d)
 }
 
-func InExpo(t, b, c, d float32) float32 {
+func InExpo(t, b, c, d float64) float64 {
 	if t == 0 {
 		return b
 	}
 	return c*pow(2, 10*(t/d-1)) + b - c*0.001
 }
 
-func OutExpo(t, b, c, d float32) float32 {
+func OutExpo(t, b, c, d float64) float64 {
 	if t == d {
 		return b + c
 	}
 	return c*1.001*(-pow(2, -10*t/d)+1) + b
 }
 
-func InOutExpo(t, b, c, d float32) float32 {
+func InOutExpo(t, b, c, d float64) float64 {
 	if t == 0 {
 		return b
 	}
@@ -158,22 +158,22 @@ func InOutExpo(t, b, c, d float32) float32 {
 	return c/2*1.0005*(-pow(2, -10*(t-1))+2) + b
 }
 
-func OutInExpo(t, b, c, d float32) float32 {
+func OutInExpo(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutExpo(t*2, b, c/2, d)
 	}
 	return InExpo((t*2)-d, b+c/2, c/2, d)
 }
 
-func InCirc(t, b, c, d float32) float32 {
+func InCirc(t, b, c, d float64) float64 {
 	return (-c*(sqrt(1-pow(t/d, 2))-1) + b)
 }
 
-func OutCirc(t, b, c, d float32) float32 {
+func OutCirc(t, b, c, d float64) float64 {
 	return (c*sqrt(1-pow(t/d-1, 2)) + b)
 }
 
-func InOutCirc(t, b, c, d float32) float32 {
+func InOutCirc(t, b, c, d float64) float64 {
 	t = t / d * 2
 	if t < 1 {
 		return -c/2*(sqrt(1-t*t)-1) + b
@@ -182,14 +182,14 @@ func InOutCirc(t, b, c, d float32) float32 {
 	return c/2*(sqrt(1-t*t)+1) + b
 }
 
-func OutInCirc(t, b, c, d float32) float32 {
+func OutInCirc(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutCirc(t*2, b, c/2, d)
 	}
 	return InCirc((t*2)-d, b+c/2, c/2, d)
 }
 
-func InElastic(t, b, c, d float32) float32 {
+func InElastic(t, b, c, d float64) float64 {
 	if t == 0 {
 		return b
 	}
@@ -202,7 +202,7 @@ func InElastic(t, b, c, d float32) float32 {
 	return -(a * pow(2, 10*t) * sin((t*d-s)*(2*pi)/p)) + b
 }
 
-func OutElastic(t, b, c, d float32) float32 {
+func OutElastic(t, b, c, d float64) float64 {
 	if t == 0 {
 		return b
 	}
@@ -214,7 +214,7 @@ func OutElastic(t, b, c, d float32) float32 {
 	return a*pow(2, -10*t)*sin((t*d-s)*(2*pi)/p) + c + b
 }
 
-func InOutElastic(t, b, c, d float32) float32 {
+func InOutElastic(t, b, c, d float64) float64 {
 	if t == 0 {
 		return b
 	}
@@ -230,24 +230,24 @@ func InOutElastic(t, b, c, d float32) float32 {
 	return a*pow(2, -10*t)*sin((t*d-s)*(2*pi)/p)*0.5 + c + b
 }
 
-func OutInElastic(t, b, c, d float32) float32 {
+func OutInElastic(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutElastic(t*2, b, c/2, d)
 	}
 	return InElastic((t*2)-d, b+c/2, c/2, d)
 }
 
-func InBack(t, b, c, d float32) float32 {
+func InBack(t, b, c, d float64) float64 {
 	t = t / d
 	return c*t*t*((backS+1)*t-backS) + b
 }
 
-func OutBack(t, b, c, d float32) float32 {
+func OutBack(t, b, c, d float64) float64 {
 	t = t/d - 1
 	return c*(t*t*((backS+1)*t+backS)+1) + b
 }
 
-func InOutBack(t, b, c, d float32) float32 {
+func InOutBack(t, b, c, d float64) float64 {
 	s := backS * 1.525
 	t = t / d * 2
 	if t < 1 {
@@ -257,14 +257,14 @@ func InOutBack(t, b, c, d float32) float32 {
 	return c/2*(t*t*((s+1)*t+s)+2) + b
 }
 
-func OutInBack(t, b, c, d float32) float32 {
+func OutInBack(t, b, c, d float64) float64 {
 	if t < (d / 2) {
 		return OutBack(t*2, b, c/2, d)
 	}
 	return InBack((t*2)-d, b+c/2, c/2, d)
 }
 
-func OutBounce(t, b, c, d float32) float32 {
+func OutBounce(t, b, c, d float64) float64 {
 	t = t / d
 	if t < 1/2.75 {
 		return c*(7.5625*t*t) + b
@@ -280,41 +280,41 @@ func OutBounce(t, b, c, d float32) float32 {
 	return c*(7.5625*t*t+0.984375) + b
 }
 
-func InBounce(t, b, c, d float32) float32 {
+func InBounce(t, b, c, d float64) float64 {
 	return c - OutBounce(d-t, 0, c, d) + b
 }
 
-func InOutBounce(t, b, c, d float32) float32 {
+func InOutBounce(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return InBounce(t*2, 0, c, d)*0.5 + b
 	}
 	return OutBounce(t*2-d, 0, c, d)*0.5 + c*.5 + b
 }
 
-func OutInBounce(t, b, c, d float32) float32 {
+func OutInBounce(t, b, c, d float64) float64 {
 	if t < d/2 {
 		return OutBounce(t*2, b, c/2, d)
 	}
 	return InBounce((t*2)-d, b+c/2, c/2, d)
 }
 
-func calculatePAS(c, d float32) (p, a, s float32) {
+func calculatePAS(c, d float64) (p, a, s float64) {
 	p = d * 0.3
 	return p, c, p / 4
 }
 
-func pow(x, y float32) float32 {
-	return float32(math.Pow(float64(x), float64(y)))
+func pow(x, y float64) float64 {
+	return float64(math.Pow(float64(x), float64(y)))
 }
 
-func cos(x float32) float32 {
-	return float32(math.Cos(float64(x)))
+func cos(x float64) float64 {
+	return float64(math.Cos(float64(x)))
 }
 
-func sin(x float32) float32 {
-	return float32(math.Sin(float64(x)))
+func sin(x float64) float64 {
+	return float64(math.Sin(float64(x)))
 }
 
-func sqrt(x float32) float32 {
-	return float32(math.Sqrt(float64(x)))
+func sqrt(x float64) float64 {
+	return float64(math.Sqrt(float64(x)))
 }
