+++ gween_test.go
@@ -11,34 +11,34 @@ import (
 func TestNew(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 
-	assert.Equal(t, float32(0), tween.begin)
-	assert.Equal(t, float32(10), tween.end)
-	assert.Equal(t, float32(10), tween.change)
-	assert.Equal(t, float32(10), tween.duration)
-	assert.Equal(t, float32(0), tween.time)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(0), tween.begin)
+	assert.Equal(t, float64(10), tween.end)
+	assert.Equal(t, float64(10), tween.change)
+	assert.Equal(t, float64(10), tween.duration)
+	assert.Equal(t, float64(0), tween.time)
+	assert.Equal(t, float64(0), tween.Overflow)
 	assert.False(t, tween.reverse)
 }
 
 func TestTween_Set(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	current, isFinished := tween.Set(2)
-	assert.Equal(t, float32(2), current)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(2), current)
+	assert.Equal(t, float64(0), tween.Overflow)
 	assert.False(t, isFinished)
 	current, isFinished = tween.Set(11)
-	assert.Equal(t, float32(10), current)
-	assert.Equal(t, float32(1), tween.Overflow)
+	assert.Equal(t, float64(10), current)
+	assert.Equal(t, float64(1), tween.Overflow)
 	assert.True(t, isFinished)
 }
 
 func TestTween_SetNeg(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	current, isFinished := tween.Set(2)
-	assert.Equal(t, float32(2), current)
+	assert.Equal(t, float64(2), current)
 	assert.False(t, isFinished)
 	current, isFinished = tween.Set(-1)
-	assert.Equal(t, float32(0), current)
+	assert.Equal(t, float64(0), current)
 	assert.False(t, isFinished)
 }
 
@@ -46,12 +46,12 @@ func TestTween_SetReverse(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	tween.reverse = true
 	current, isFinished := tween.Set(2)
-	assert.Equal(t, float32(2), current)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(2), current)
+	assert.Equal(t, float64(0), tween.Overflow)
 	assert.False(t, isFinished)
 	current, isFinished = tween.Set(11)
-	assert.Equal(t, float32(10), current)
-	assert.Equal(t, float32(1), tween.Overflow)
+	assert.Equal(t, float64(10), current)
+	assert.Equal(t, float64(1), tween.Overflow)
 	assert.False(t, isFinished)
 }
 
@@ -59,23 +59,23 @@ func TestTween_SetNegReverse(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	tween.reverse = true
 	current, isFinished := tween.Set(2)
-	assert.Equal(t, float32(2), current)
+	assert.Equal(t, float64(2), current)
 	assert.False(t, isFinished)
 	current, isFinished = tween.Set(-1)
-	assert.Equal(t, float32(0), current)
+	assert.Equal(t, float64(0), current)
 	assert.True(t, isFinished)
 }
 
 func TestTween_Reset(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	current, isFinished := tween.Set(2)
-	assert.Equal(t, float32(2), current)
-	assert.Equal(t, float32(2), tween.time)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(2), current)
+	assert.Equal(t, float64(2), tween.time)
+	assert.Equal(t, float64(0), tween.Overflow)
 	assert.False(t, isFinished)
 	tween.Reset()
-	assert.Equal(t, float32(0), tween.time)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(0), tween.time)
+	assert.Equal(t, float64(0), tween.Overflow)
 }
 
 func TestTween_ResetReverse(t *testing.T) {
@@ -83,19 +83,19 @@ func TestTween_ResetReverse(t *testing.T) {
 	tween.Set(2)
 	tween.reverse = true
 	tween.Reset()
-	assert.Equal(t, float32(10), tween.time)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(10), tween.time)
+	assert.Equal(t, float64(0), tween.Overflow)
 }
 
 func TestTween_Update(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	current, isFinished := tween.Update(2)
-	assert.Equal(t, float32(2), current)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(2), current)
+	assert.Equal(t, float64(0), tween.Overflow)
 	assert.False(t, isFinished)
 	current, isFinished = tween.Update(9)
-	assert.Equal(t, float32(10), current)
-	assert.Equal(t, float32(1), tween.Overflow)
+	assert.Equal(t, float64(10), current)
+	assert.Equal(t, float64(1), tween.Overflow)
 	assert.True(t, isFinished)
 }
 
@@ -103,7 +103,7 @@ func TestTween_UpdateZero(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	tween.Update(2)
 	current, isFinished := tween.Update(0)
-	assert.Equal(t, float32(2), current)
+	assert.Equal(t, float64(2), current)
 	assert.False(t, isFinished)
 }
 
@@ -111,7 +111,7 @@ func TestTween_UpdateNeg(t *testing.T) {
 	tween := New(0, 10, 10, ease.Linear)
 	tween.Update(2)
 	current, isFinished := tween.Update(-1)
-	assert.Equal(t, float32(1), current)
+	assert.Equal(t, float64(1), current)
 	assert.False(t, isFinished)
 }
 
@@ -125,7 +125,7 @@ func TestTween_CanReverse(t *testing.T) {
 	tween.Update(8)
 	tween.reverse = true
 	current, isFinished := tween.Update(2)
-	assert.Equal(t, float32(6), current)
+	assert.Equal(t, float64(6), current)
 	assert.False(t, isFinished)
 }
 
@@ -135,7 +135,7 @@ func TestTween_CanReverseFromFinished(t *testing.T) {
 	assert.True(t, isFinished)
 	tween.reverse = true
 	current, isFinished = tween.Update(2)
-	assert.Equal(t, float32(8), current)
+	assert.Equal(t, float64(8), current)
 	assert.False(t, isFinished)
 }
 
@@ -144,10 +144,10 @@ func TestTween_CanReverseFromStart(t *testing.T) {
 	tween.reverse = true
 	current, isFinished := tween.Update(0)
 	assert.True(t, isFinished)
-	assert.Equal(t, float32(0), current)
-	assert.Equal(t, float32(0), tween.Overflow)
+	assert.Equal(t, float64(0), current)
+	assert.Equal(t, float64(0), tween.Overflow)
 	current, isFinished = tween.Update(1)
 	assert.True(t, isFinished)
-	assert.Equal(t, float32(0), current)
-	assert.Equal(t, float32(-1.0), tween.Overflow)
+	assert.Equal(t, float64(0), current)
+	assert.Equal(t, float64(-1.0), tween.Overflow)
 }
